LayoutTests/fast/forms/month/ValidityState-stepMismatch-month-expected.txt

CONSOLE MESSAGE: line 55: SyntaxError: Unexpected token '<'
